# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and return token",
                "operationId": "ed510e353bad3b3e79011fe97bc090cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "some-generated-token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user by deleting current access token",
                "operationId": "ba942f3ef9bd9df4eb886a04dd50f2a7",
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "logout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout-all": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user from all devices by deleting all access tokens",
                "operationId": "6eac3154f743e12de6d1c0c4318b8e52",
                "responses": {
                    "200": {
                        "description": "Successful absolute logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "logout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/portfolio/index": {
            "get": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Get user portfolio",
                "description": "Get the portfolio of the authenticated user",
                "operationId": "2d01b406e1ab957cab24814f8adac4d9",
                "responses": {
                    "200": {
                        "description": "A list of portfolio items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PortfolioResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/portfolio/show/{id}": {
            "get": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Get portfolio by ID",
                "description": "Retrieve a specific portfolio by its ID",
                "operationId": "dd53822bdfe179fb708d0e8dd89181a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the portfolio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested portfolio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PortfolioResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Portfolio not found"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/portfolio/store": {
            "post": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Create a new portfolio",
                "description": "Create a new portfolio for the authenticated user",
                "operationId": "6c14ca22ab7e4256030fe3b6832c8e31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the portfolio",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Portfolio created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PortfolioResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/portfolio/update/{id}": {
            "patch": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Update an existing portfolio",
                "description": "Update the details of a portfolio",
                "operationId": "af330cb990fee95402f991be8551d975",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the portfolio to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the portfolio",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Portfolio updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PortfolioResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Portfolio not found"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/portfolio/delete/{id}": {
            "delete": {
                "tags": [
                    "Portfolio"
                ],
                "summary": "Delete a portfolio",
                "description": "Delete a specific portfolio by ID",
                "operationId": "6169795749018ff6280730aea06a6972",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the portfolio to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Portfolio deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Portfolio not found"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PortfolioResource": {
                "properties": {
                    "id": {
                        "description": "ID of the portfolio",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the portfolio",
                        "type": "string"
                    },
                    "balance": {
                        "description": "Portfolio balance in USD as calculated from transactions",
                        "type": "number",
                        "format": "decimal"
                    },
                    "created_at": {
                        "description": "Portfolio creation time",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated": {
                        "description": "Portfolio last update time",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "SanctumAuth": {
                "type": "apiKey",
                "description": "Введите токен в формате 'Bearer {your-token}'",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}